Added ownership check for photo management

+++ Systems/API/ZooExpoOrg.Api/Controllers/Photos/PhotoController.cs
@@ -8,6 +8,7 @@ using ZooExpoOrg.Common.Exceptions;
 using ZooExpoOrg.Common.Security;
 using ZooExpoOrg.Services.Logger;
 using ZooExpoOrg.Services.Photos;
+using ZooExpoOrg.Services.RightVerifier;
 
 [ApiController]
 [ApiVersion("1.0")]
@@ -18,12 +19,18 @@ public class PhotoController : ControllerBase
     private readonly IAppLogger logger;
     private readonly IPhotoService photoService;
     private readonly IMapper mapper;
+    private readonly IRightVerifierService rightVerifier;
 
-    public PhotoController(IAppLogger logger, IPhotoService photoService, IMapper mapper)
+    public PhotoController(
+        IAppLogger logger, 
+        IPhotoService photoService, 
+        IMapper mapper,
+        IRightVerifierService rightVerifier)
     {
         this.logger = logger;
         this.photoService = photoService;
         this.mapper = mapper;
+        this.rightVerifier = rightVerifier;
     }
 
     [HttpGet("owned/{ownerId:Guid}")]
@@ -50,11 +57,18 @@ public class PhotoController : ControllerBase
 
     [HttpPost("")]
     [Authorize(AppScopes.UseScope)]
-    public async Task<IActionResult> Create(CreatePhotoModel createModel)
+    public async Task<IActionResult> Create(CreatePhotoModel model)
     {
         try
         {
-            var result = await photoService.Create(createModel);
+            string jwtToken = Request.Headers["Authorization"].ToString().Replace("Bearer ", "");
+
+            if (!(await rightVerifier.VerifRightsOfCreatePhoto(jwtToken, model.OwnerId, model.LocationId)))
+            {
+                return BadRequest("Access denied.");
+            }
+
+            var result = await photoService.Create(model);
 
             return Ok(mapper.Map<PresintationPhotoModel>(result));
         }
@@ -70,6 +84,13 @@ public class PhotoController : ControllerBase
     {
         try
         {
+            string jwtToken = Request.Headers["Authorization"].ToString().Replace("Bearer ", "");
+
+            if (!(await rightVerifier.VerifRightsOfManagPhoto(jwtToken, id)))
+            {
+                return BadRequest("Access denied.");
+            }
+
             await photoService.Delete(id);
 
             return Ok();
